Rename to StatueRoutes
