+++ src/components/index.js
@@ -2,4 +2,4 @@ export { StatueMenuContainer } from "./StatueMenuContainer";
 export { Hero } from "./Hero.js";
 export { NavContainer } from "./NavContainer";
 export { StatueDetailContainer } from "./StatueDetailContainer";
-export { StatueRoute } from "./StatueRoute";
+export { StatueRoutes } from "./StatueRoutes";
